+++ TabletopTweaks/NewContent/ContentAdder.cs
@@ -75,6 +75,8 @@ namespace TabletopTweaks.NewContent {
                 MythicAbilities.MythicSpellCombat.AddMythicSpellCombat();
                 MythicAbilities.PrecisionCritical.AddPrecisionCritical();
 
+                MythicFeats.MythicShatterDefenses.AddMythicShatterDefenses();
+
                 Feats.MagicalAptitude.AddMagicalAptitude();
                 Feats.Scholar.AddScholar();
                 Feats.SelfSufficient.AddSelfSufficient();
@@ -82,6 +84,10 @@ namespace TabletopTweaks.NewContent {
                 Feats.StreetSmarts.AddStreetSmarts();
                 Feats.GracefulAthlete.AddGracefulAthlete();
                 Feats.DervishDance.AddDervishDance();
+                Feats.NatureSoul.AddNatureSoul();
+                Feats.AnimalAlly.AddAnimalAlly();
+
+                Feats.ShatterDefenses.AddNewShatterDefenseBlueprints();
 
                 Feats.ProdigiousTWF.AddProdigiousTWF();
 
@@ -94,7 +100,7 @@ namespace TabletopTweaks.NewContent {
                 Feats.ExtraSlayerTalent.AddExtraSlayerTalent();
                 Feats.ExtraRevelation.AddExtraRevelation();
                 Feats.ExtraDiscovery.AddExtraDiscovery();
-                Feats.ExtraMercy.AddExtraMercy();
+                Feats.ExtraMercy.AddExtraMercy();             
 
                 AlternateCapstones.MasterfulTalent.AddMasterfulTalent();